// FrontEnd/src/app/api/product.ts
export interface Product {
    id?: string;
    code?: string;
    name?: string;
    description?: string;
    price?: number;
    quantity?: number;
    inventoryStatus?: string;
    category?: string;
    image?: string;
    rating?: number;

    obatalkes_id?: string;
    obatalkes_nama?: string;
    imasigna_id?: string;
    signa_nama?: string;
    qty?: number;
    nama_id?: string;
    
}